emulator: Add a fourth element in exception stacktraces
This commit is a preparation for introducing location information
(filename/line number) in stacktraces in exceptions. Currently
a stack trace looks like:

  [{Mod1,Function1,Arity1},
   .
   .
   .
   {ModN,FunctionN,ArityN}]

Add a forth element to each tuple that can be used indication
the filename and line number of the source file:

  [{Mod1,Function1,Arity1,Location1},
   .
   .
   .
   {ModN,FunctionN,ArityN,LocationN}]

In this commit, the fourth element will just be an empty list,
and we will change all code that look at or manipulate stacktraces.